+++ helianto-core/src/main/java/org/helianto/core/security/PublicUserDetails.java
@@ -44,7 +44,7 @@ public interface PublicUserDetails {
     /**
      * Identity security id.
      */
-    long getIdentitySecurityId();
+    int getIdentitySecurityId();
     
     
 }
