+++ helianto-core/src/main/java/org/helianto/core/security/UserDetailsAdapter.java
@@ -110,7 +110,7 @@ public class UserDetailsAdapter
     /**
      * Identity security id.
      */
-    public long getIdentitySecurityId() {
+    public int getIdentitySecurityId() {
     	if (identitySecurity!=null) {
     		return identitySecurity.getId();
     	}
